+++ src/test/java/com/example/budongbudong/fixture/AuctionFixture.java
@@ -4,6 +4,7 @@ import com.example.budongbudong.common.entity.Auction;
 import com.example.budongbudong.common.entity.Property;
 import com.example.budongbudong.domain.auction.enums.AuctionStatus;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public class AuctionFixture {
@@ -14,7 +15,7 @@ public class AuctionFixture {
     ) {
         return Auction.testBuilder()
                 .property(property)
-                .startPrice(1000L)
+                .startPrice(BigDecimal.valueOf(1000L))
                 .status(AuctionStatus.OPEN)
                 .startedAt(startOfToday.minusDays(2))
                 .endedAt(startOfToday.minusSeconds(1))
@@ -27,7 +28,7 @@ public class AuctionFixture {
     ) {
         return Auction.testBuilder()
                 .property(property)
-                .startPrice(1000L)
+                .startPrice(BigDecimal.valueOf(1000L))
                 .status(AuctionStatus.OPEN)
                 .startedAt(startOfToday.minusDays(1))
                 .endedAt(startOfToday.plusDays(1))
